fin documentacion inicio test

+++ src/main/java/edu/escuelaing/arep/LinkedList.java
@@ -1,14 +1,25 @@
 package edu.escuelaing.arep;
-
+/**
+ * implementacion de una linked list simple 
+ * @author David Ramirez
+ */
 public class LinkedList{
     private static Nodo head;
     public LinkedList(){
         head=null;
     }
+    /**
+     * metodo para obtener el item que se encuentra en la primera posicion de la linked list
+     * @return head el primero elemento de la linked list
+     */
     public Nodo getHead(){
         return head;
     }
-    public  void insertar(int value){
+    /**
+     * metodo para insertar un elemento al final de la linked list
+     * @param value el valor del elemento a insertar en la linked list
+     */
+    public  void insertar(float value){
         if (head==null){
             head= new Nodo(null,value);
         }
@@ -22,7 +33,11 @@ public class LinkedList{
         }
         System.out.println("valor insertado");
     }
-    public  void eliminar(int value){
+    /**
+     * metodo para eliminar un elemento de la linked list 
+     * @param value el valor del elemento a eliminar de la linked list
+     */
+    public  void eliminar(float value){
         boolean eliminado=false;
         if(head!=null){
             if(head.value()==value){
@@ -50,6 +65,9 @@ public class LinkedList{
             System.out.println("no se encontro el elemento");
         }
     }
+    /**
+     * metodo que imprime en pantalla los datos contenidos en la linked list
+     */
     public  void verDatos(){
         Nodo actual=head;
         while(actual!=null){
